Document, restructure, and refactor core
