+++ packages/core/scripts/test-examples.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
+# Validate the example packages used for the tests by compiling and running them.
+
 set -e
 shopt -s globstar
