+++ packages/core/src/dowdep.ts
@@ -1,18 +1,15 @@
-import pathExists from 'path-exists'
 import path from 'path'
+import pathExists from 'path-exists'
 
 import { createDependencySearcher, Dependency, DependencySearcher, DependencySearchStrategy } from './dependencies'
 import { Package } from './packages'
-import { PackageReferenceSearcher, ReferenceSearchStrategy } from './references'
+import { createReferenceSearcher, ReferenceSearchStrategy } from './references'
 import { OnlyData } from './utils/OnlyData'
 
 
-export function getCacheDirectory() {
-    return process.env.NPM_CACHE || 'cache'
-}
-
+/** This is the starting point and configuration place for the downstream repository miner. */
 export class Dowdep {
-    constructor(init: Partial<OnlyData<Dowdep>>) {
+    constructor(init: Partial<OnlyData<Dowdep>> = {}) {
         Object.assign(this, init)
     }
 
@@ -40,17 +37,18 @@ export class Dowdep {
             ? <['npm', 'sourcegraph']>['npm', 'sourcegraph']
             : this.dependencySearchStrategies
 
-        return strategies.map(strategy => createDependencySearcher(strategy, $package, {
+        return strategies.map(strategy => createDependencySearcher($package, strategy, {
             limit: this.dependencyLimit ? Math.ceil(this.dependencyLimit / strategies.length) : this.dependencyLimit
         }))
     }
 
     createReferenceSearcher(dependency: Dependency, $package: Package) {
         // TODO: Do we still need $package param?
-        return PackageReferenceSearcher.create($package, dependency, this.referenceSearchStrategy)
+        return createReferenceSearcher($package, dependency, this.referenceSearchStrategy)
     }
 }
 
+/** Wraps an existing physical or virtual filesystem. In preparation for supporting VS Code's virtual filesystem for github.dev. */
 export interface FileSystem {
     exists(path: string): Promise<boolean>
     join(...parts: string[]): string
