+++ packages/core/src/references/index.ts
@@ -1,0 +1,93 @@
+import { promises as fsPromises, Dirent } from 'fs'
+import { Dependency } from '../dependencies'
+import { Package } from '../packages'
+import path from 'path'
+
+import { ALL_REFERENCE_KINDS, Reference, ReferenceKind } from './base'
+import { HeuristicReferenceSearcher } from './heuristic'
+import { TypeReferenceSearcher } from './types'
+import fsExists from '../utils/fs.exists'
+
+export {
+    DeclarationLocation,
+    FilePosition,
+    FilePositionPrimitive,
+    Location,
+    Reference,
+    ReferenceKind,
+    ReferenceSearcher
+} from './base'
+
+
+const ALL_REFERENCE_SEARCH_STRATEGIES = [
+    'heuristic',
+    'types'
+] as const
+export type ReferenceSearchStrategy = (typeof ALL_REFERENCE_SEARCH_STRATEGIES)[number]
+
+function referenceSearcher(strategy: ReferenceSearchStrategy) {
+    switch (strategy) {
+        case 'heuristic':
+            return HeuristicReferenceSearcher
+        case 'types':
+            return TypeReferenceSearcher
+    }
+}
+
+export function createReferenceSearcher($package: Package, dependency: Dependency, strategy: ReferenceSearchStrategy) {
+    return new (referenceSearcher(strategy))($package, dependency)
+}
+
+/** Collects all packages from a `node_modules`-like folder structure and searches references in each package. */
+export class ReferenceCollector {
+    package: Package
+    rootDirectory: string
+    searchStrategy: ReferenceSearchStrategy
+
+    constructor($package: Package, rootDirectory: string, searchStrategy: ReferenceSearchStrategy) {
+        this.package = $package
+        this.rootDirectory = rootDirectory
+        this.searchStrategy = searchStrategy
+    }
+
+    async* searchReferences(limit?: number, includeKinds: readonly ReferenceKind[] | '*' = ['usage']): AsyncIterable<Reference> {
+        yield* this.basicSearchReferences(this.rootDirectory, limit, includeKinds == '*' ? ALL_REFERENCE_KINDS : includeKinds, 0)
+    }
+
+    protected async* basicSearchReferences(rootDirectory: string, limit: number | undefined, includeKinds: readonly ReferenceKind[] | '*', depth: number): AsyncIterable<Reference> {
+        if (!await fsExists(path.join(rootDirectory, 'package.json'))) {
+            // Search recursively
+            console.log("read", rootDirectory)
+            let depDirectories: Iterable<Dirent> = []
+            try {
+                depDirectories = (
+                    await fsPromises.readdir(rootDirectory, { withFileTypes: true })
+                ).filter(dirent => dirent.isDirectory)
+            } catch (ex) {
+                console.error(ex)
+            }
+            console.log("did read", depDirectories)
+
+            let i = 0
+            for await (const depDirectory of depDirectories) {
+                for await (const reference of this.basicSearchReferences(path.join(rootDirectory, depDirectory.name), undefined, includeKinds, depth + 1)) {
+                    yield reference
+                    if (limit && ++i >= limit) {
+                        return
+                    }
+                }
+            }
+            return
+        }
+
+        const dependencyName = path.basename(rootDirectory)
+        const searcher = createReferenceSearcher(
+            this.package,
+            new Dependency(dependencyName, new Package("hack", path.dirname(rootDirectory))),
+            this.searchStrategy
+        )
+        searcher.includeKinds = includeKinds
+        await searcher.initialize()
+        yield* searcher.searchReferences(rootDirectory)
+    }
+}
