+++ packages/core/test/references.test.ts
@@ -1,9 +1,9 @@
-import asyncIteratorToArray from 'it-all'
+import itAll from 'it-all'
 import _ from 'lodash'
 import { jsonc as json } from 'jsonc'
 
 import { Package } from '../src/packages'
-import { ReferenceSearcher, ReferenceKind } from '../src/references'
+import { ReferenceCollector, ReferenceKind, ReferenceSearchStrategy } from '../src/references'
 import ifCurtailed from '../src/utils/if-curtailed'
 import { lodashClassifyNested } from '../src/utils/lodash-classify'
 import { getSourceDirectory } from './_utils/sourceDirectory'
@@ -32,19 +32,19 @@ const expectedTypeReferences = <References>json.readSync(`${SOURCE_DIRECTORY}/ex
 describe('ReferenceSearcher', () => {
     it.each(_.flatMap(
         Object.entries({'heuristic': expectedHeuristicReferences, 'types': expectedTypeReferences}),
-        ([packageReferenceSearcher, allExpectedReferences]) => _.map(
+        ([strategy, allExpectedReferences]) => _.map(
             Object.entries(allExpectedReferences),
-            ([packageName, expectedReferences]) => ({ packageReferenceSearcher, packageName, expectedReferences }))
+            ([packageName, expectedReferences]) => ({ strategy: <ReferenceSearchStrategy>strategy, packageName, expectedReferences }))
     ))("should find relevant references for %s", async (
-        { packageReferenceSearcher, packageName, expectedReferences }) => {
+        { strategy, packageName, expectedReferences }) => {
         const $package = new Package(
             packageName,
             `${SOURCE_DIRECTORY}/examples/packages/${packageName}`
         )
-        const searcher = new ReferenceSearcher($package, `${SOURCE_DIRECTORY}/examples/dependents`, packageReferenceSearcher)
-        const references = await asyncIteratorToArray(searcher.searchReferences(undefined, '*'))
+        const collector = new ReferenceCollector($package, `${SOURCE_DIRECTORY}/examples/dependents`, strategy)
+        const references = await itAll(collector.searchReferences(undefined, '*'))
 
-        /** Since null and undefined are invalid keys in JS objects, we stringify them for compatibility with lodash. See Reference.memberName. */
+        /** Since null and undefined are invalid keys in JS objects, we stringify them for compatibility with lodash dictionaries. See Reference.memberName. */
         function stringify(key: string | null | undefined) {
             if (key === undefined) return '<undefined>'
             if (key === null) return '<null>'
@@ -66,7 +66,7 @@ describe('ReferenceSearcher', () => {
             .value()
 
         // Classify actual-expected data by whether we want to do make exact assertions or not
-        const isExactExpectation = (kind: ReferenceKind) => packageReferenceSearcher != 'heuristic' && kind != 'occurence'
+        const isExactExpectation = (kind: ReferenceKind) => strategy != 'heuristic' && kind != 'occurence'
         const allExpectations = lodashClassifyNested(
             {
                 actual: aggregatedReferences,
@@ -92,7 +92,7 @@ describe('ReferenceSearcher', () => {
                 )))
             ifCurtailed(
                 () => expect(actual).toEqual(containingExpected),
-                () => printDiff(actual, containingExpected, packageReferenceSearcher, packageName))
+                () => printDiff(actual, containingExpected, strategy, packageName))
         })
     })
 })
