+++ packages/example-js/index.js
@@ -2,11 +2,10 @@
 
 const dowdep = require('dowdep')
 
-const cacheDirectory = dowdep.getCacheDirectory()
 const $package = new dowdep.Package(
     'name',
     '/path/to/directory'
 )
-console.log(cacheDirectory)
 console.log($package)
+console.log(new dowdep.Dowdep().createDependencySearchers($package))
 console.log(dowdep.ReferenceSearcher)
